fix banner types causing auth error
