+++ src/data/queries/event.ts
@@ -39,10 +39,6 @@ export const data: QueryData<EventDataOpts, NodeEvent> = async (
         opts.context,
         {
           params: params().getQueryObject(),
-          // withAuth: {
-          //   clientId: process.env.DRUPAL_CLIENT_ID,
-          //   clientSecret: process.env.DRUPAL_CLIENT_SECRET,
-          // }
         }
       )
     : // otherwise just lookup by uuid
