+++ src/data/queries/newsStory.ts
@@ -36,10 +36,6 @@ export const data: QueryData<NewsStoryDataOpts, NodeNewsStory> = async (
         opts.context,
         {
           params: params().getQueryObject(),
-          // withAuth: {
-          //   clientId: process.env.DRUPAL_CLIENT_ID,
-          //   clientSecret: process.env.DRUPAL_CLIENT_SECRET,
-          // },
         }
       )
     : // otherwise just lookup by uuid
