+++ src/lib/drupal/drupalClient.ts
@@ -9,10 +9,10 @@ export const drupalClient = new DrupalClient(baseUrl, {
   fetcher: getFetcher(baseUrl),
   useDefaultResourceTypeEntry: true,
   throwJsonApiErrors: false,
-  // auth: {
-  //   clientId: process.env.DRUPAL_CLIENT_ID,
-  //   clientSecret: process.env.DRUPAL_CLIENT_SECRET,
-  // },
+  auth: {
+    clientId: process.env.DRUPAL_CLIENT_ID,
+    clientSecret: process.env.DRUPAL_CLIENT_SECRET,
+  },
   // cache: redisCache(createRedisClient(process.env.REDIS_URL)),
   previewSecret: process.env.DRUPAL_PREVIEW_SECRET,
 })
